+++ src/features/home/components/timer/getUserStats.ts
@@ -1,9 +1,8 @@
 import { User } from "@firebase/auth"
-import { firestore } from "features/home/firebase-init"
-import { doc, getDoc } from "firebase/firestore"
+import { Firestore, doc, getDoc } from "firebase/firestore"
 
-export const getUserStats = async (user: User) => {
-  const userStatsRef = doc(firestore, "users", user.uid)
+export const getUserStats = async (user: User, firestoreDB: Firestore) => {
+  const userStatsRef = doc(firestoreDB, "users", user.uid)
   const userStatsDoc = await getDoc(userStatsRef)
   const userStatsData = userStatsDoc.data()
   return userStatsData